combined gold feature store processing in one script, updated data pipeline
