+++ assignment_2/dags/dag_data_pipeline.py
@@ -74,12 +74,7 @@ with DAG(
  
     # --- feature store ---
 
-    # check if source for clickstream data is available
-    dep_check_source_clickstream_data = PythonOperator(
-        task_id='dep_check_source_clickstream_data',
-        python_callable=lambda: print("Checking if source clickstream data is available...")
-        # This is a placeholder; replace with actual check logic if needed
-    )
+    ### Customer Attributes Processing
 
     # check if source for customer attributes data is available
     dep_check_source_customer_attributes_data = PythonOperator(
@@ -88,23 +83,6 @@ with DAG(
         # This is a placeholder; replace with actual check logic if needed
     )
 
-    # check if source for customer financial data is available
-    dep_check_source_customer_financial_data = PythonOperator(
-        task_id='dep_check_source_customer_financial_data',
-        python_callable=lambda: print("Checking if source customer financial data is available...")
-        # This is a placeholder; replace with actual check logic if needed
-    )
-
-    # run bronze clickstream store script
-    bronze_clickstream_store = BashOperator(
-        task_id='run_bronze_clickstream_store',
-        bash_command=(
-            'cd /opt/airflow/scripts/data_processing && '
-            'python bronze_processing.py '
-            '--date "{{ ds }}" --type clickstream_data'
-        ),
-    )
-    
     # run bronze customer attributes store script
     bronze_customer_attributes_store = BashOperator(
         task_id='run_bronze_customer_attributes_store',
@@ -112,26 +90,6 @@ with DAG(
             'cd /opt/airflow/scripts/data_processing && '
             'python bronze_processing.py '
             '--date "{{ ds }}" --type customer_attributes'
-        ),
-    )
-
-    # run bronze customer financials store script
-    bronze_customer_financials_store = BashOperator(
-        task_id='run_bronze_customer_financials_store',
-        bash_command=(
-            'cd /opt/airflow/scripts/data_processing && '
-            'python bronze_processing.py '
-            '--date "{{ ds }}" --type customer_financials'
-        ),
-    )
-
-    # run silver clickstream store script
-    silver_clickstream_store = BashOperator(
-        task_id='run_silver_clickstream_store',
-        bash_command=(
-            'cd /opt/airflow/scripts/data_processing && '
-            'python silver_processing.py '
-            '--date "{{ ds }}" --type clickstream_data'
         ),
     )
 
@@ -145,6 +103,25 @@ with DAG(
         ),
     )
 
+    ### Customer Financials Processing
+
+    # check if source for customer financial data is available
+    dep_check_source_customer_financial_data = PythonOperator(
+        task_id='dep_check_source_customer_financial_data',
+        python_callable=lambda: print("Checking if source customer financial data is available...")
+        # This is a placeholder; replace with actual check logic if needed
+    )
+
+    # run bronze customer financials store script
+    bronze_customer_financials_store = BashOperator(
+        task_id='run_bronze_customer_financials_store',
+        bash_command=(
+            'cd /opt/airflow/scripts/data_processing && '
+            'python bronze_processing.py '
+            '--date "{{ ds }}" --type customer_financials'
+        ),
+    )
+
     # run silver customer financials store script
     silver_customer_financials_store = BashOperator(
         task_id='run_silver_customer_financials_store',
@@ -155,23 +132,50 @@ with DAG(
         ),
     )
 
-    # join silver feature data to gold feature store
-    gold_feature_store_join = BashOperator(
-        task_id='run_join_gold_feature_store',
+    ### Clickstream Data Processing
+
+    # check if source for clickstream data is available
+    dep_check_source_clickstream_data = PythonOperator(
+        task_id='dep_check_source_clickstream_data',
+        python_callable=lambda: print("Checking if source clickstream data is available...")
+        # This is a placeholder; replace with actual check logic if needed
+    )
+
+    # run bronze clickstream store script
+    bronze_clickstream_store = BashOperator(
+        task_id='run_bronze_clickstream_store',
         bash_command=(
             'cd /opt/airflow/scripts/data_processing && '
-            'python gold_processing_feature_join.py '
-            '--date "{{ ds }}"'
+            'python bronze_processing.py '
+            '--date "{{ ds }}" --type clickstream_data'
+        ),
+    )
+    
+    # run silver clickstream store script
+    silver_clickstream_store = BashOperator(
+        task_id='run_silver_clickstream_store',
+        bash_command=(
+            'cd /opt/airflow/scripts/data_processing && '
+            'python silver_processing.py '
+            '--date "{{ ds }}" --type clickstream_data'
         ),
     )
 
-    # Gold Feature store data engineering
-    gold_feature_store_engineered = EmptyOperator(task_id="run_engineer_gold_feature_store")
+    ### Gold Feature Store Processing
+
+    # join silver feature data to gold feature store & engineer gold features for ml modeling
+    gold_feature_store = BashOperator(
+        task_id='run_gold_feature_store',
+        bash_command=(
+            'cd /opt/airflow/scripts/data_processing && '
+            'python gold_processing_features.py --date "{{ ds }}"'
+        ),
+    )
 
     feature_store_completed = EmptyOperator(task_id="feature_store_completed")
     
     # Define task dependencies to run scripts sequentially
-    dep_check_source_clickstream_data >> bronze_clickstream_store >> silver_clickstream_store >> gold_feature_store_join # type: ignore
-    dep_check_source_customer_attributes_data >> bronze_customer_attributes_store >> silver_customer_attributes_store >> gold_feature_store_join # type: ignore
-    dep_check_source_customer_financial_data >> bronze_customer_financials_store >> silver_customer_financials_store >> gold_feature_store_join # type: ignore
-    gold_feature_store_join >> gold_feature_store_engineered >> feature_store_completed # type: ignore
+    dep_check_source_clickstream_data >> bronze_clickstream_store >> silver_clickstream_store >> gold_feature_store # type: ignore
+    dep_check_source_customer_attributes_data >> bronze_customer_attributes_store >> silver_customer_attributes_store >> gold_feature_store # type: ignore
+    dep_check_source_customer_financial_data >> bronze_customer_financials_store >> silver_customer_financials_store >> gold_feature_store # type: ignore
+    gold_feature_store >> feature_store_completed # type: ignore
